Event properties and HttpConsumer added
Path values (previously known only as "events") can now be arrays.
Path arrays must define a "run" key, which must contain and event.
Other keys of the path array will be consumed as properties.

The first property of this kind is "http-consumer", which allows the event (the  function stored inside the "run" key of the array) to yeild execution with an object of type HttpConsumer.

This object will allow you to read contents from the http body as the server is reading them live.

HttpConsumer::hasMore() will return true if there's more data to be read.

HttpConsumer::consume(&$output) will write the current package of data read by the server to the $output string.

HttpConsumer::rewind() has no effect except setting the inner $first variable to true which as of now does nothing.