HttpResponse fix
Fixed an issue with HttpResponse which cause the server to skip the parsing of the user "Accept" header and always return a json object.

The server should now correctly conver all responses to json,txt,or xml as requested by the user regardless of what the data on the server looks like.