Yielding functions
It is now possible to yield functions plain functions as if they were Promises.
Yielding functions (which are Generators) will be converted into Promises and they can be nested.

A yielding Closure can yield a Promise, but a Promise cannot yield a Closure, that is because the Promise implementation of ReactPHP does not recognize such a thing.

Promises can, however, resolve more promises and those will be recognized by the engine and executed in the main loop.